#44 moved TableTest to api.TableTest
